Promote cert types, loader, and classifier into shared package
The cert primitives (DevCert/DevKey wrappers, PFX I/O, PEM/PFX loader)
plus the dev-cert validation + classification helpers move from the UI
extension's src/cert/ tree into src/shared/src/cert/ so the workspace
extension can use the same code path when scanning the container for a
dev cert to push back to the host (preparation for issue #63 / #62
opt-in container-to-host sync).

The classifier becomes side-effect-free: callers in vscode contexts
opt into a localized log line via an onSkipped / onMultipleCandidates
callback, keeping vscode.l10n out of shared. The host's existing
classifyCandidate / selectBestDevCert wrappers in platform/baseStore
preserve the previous log behavior verbatim by routing the new reason
codes through the same l10n strings, so the host's test suite continues
to pass without modification.

The workspace extension now needs the reflect-metadata polyfill (loaded
transitively via @peculiar/x509 → tsyringe); added at the entry point
and in the vitest setup.
