+++ src/vscode-workspace-extension/tsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "module": "Node16",
     "target": "ES2022",
-    "lib": ["ES2022"],
+    "lib": ["ES2022", "DOM"],
     "moduleResolution": "Node16",
     "types": ["node"],
     "outDir": "dist",
